requirements.txt
openai
PyPDF2
faiss-cpu
nltk
transformers